- PostRepository에서 addTrack() 메소드가 작동하지 않는 버그 수정
- Category ID타입을 String(varchar)에서 long(bigint)로 변경

+++ trunk/blog/src/main/java/kr/co/webcash/repository/CategoryRepository.java
@@ -18,6 +18,6 @@ public interface CategoryRepository {
 
 	Category findByIdAndBlogId(String id, String blogId);
 
-	int findAllCountByBlogId(String blogId);
+	long findAllCountByBlogId(String blogId);
 	
 }
